requirements.txt
typer[all]
pymongo
python-dotenv
requests